requirements.txt
asgiref==3.8.1
click==8.2.1
Django==5.2.3
django-split-settings==1.3.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dlib==20.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==2.2.6
pillow==11.2.1
PyJWT==2.9.0
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
